Aufgabe 7 Task 1 fertig
Mein Bild ist jetzt was eigenes weil aus dem Satz:

Visualize the different confidence intervals as an image. Define the range from 150cm to 200cm. Color the parts in black in which the current confidence interval overlaps. Color the line in white for the remaining parts. Stack the lines such that they form an image build of horizontal lines. Add a gray (value 150) horizontal line at the position of the population mean of the corrsponding feature. You can use the following code to create an image

kann ich mir einfach nichts zusammen reimen.